Updated the command_executor to validate the ssl cert when using a proxy.
Update to api_spec to set the proxy_host to false otherwise we do not
want to use Net::HTTP with a proxy for these tests. Updated the tests
to account for the class attributes that set secure_mode, test_mode,
and debug_mode.

Implemented the to_s method on API::Error so that the message will
show up as part of the exception.